requirements.txt
nicegui
mysql-connector-python